fix(web): Announcement Component url 수정
