+++ apps/web/src/components/common/announcement/announcement.tsx
@@ -1,16 +1,28 @@
+import * as React from 'react';
+import { css } from 'styled-components';
+import { useRouter } from 'next/navigation';
 import {
   applyMediaQuery,
   Button,
   Typography,
   Wrapper,
 } from '@supercarmarket/ui';
-import Link from 'next/link';
-import { css } from 'styled-components';
-import * as React from 'react';
-
 import ArrowRight from '../../../assets/svg/arrow-right.svg';
 
-const MarketBanner = () => {
+interface BannerProps {
+  title: string;
+  subtitle?: string;
+  btnTitle: string;
+  url: string;
+}
+
+const Announcement = ({ title, subtitle, btnTitle, url }: BannerProps) => {
+  const { push } = useRouter();
+
+  const moveToUrl = () => {
+    push(url);
+  };
+
   return (
     <Wrapper
       css={css`
@@ -22,7 +34,6 @@ const MarketBanner = () => {
         padding: 34px 40px;
         background: ${({ theme }) => theme.color['greyScale-2']};
         border-radius: 4px;
-
         ${applyMediaQuery('mobile')} {
           flex-direction: column;
           align-items: flex-start;
@@ -38,11 +49,13 @@ const MarketBanner = () => {
         `}
       >
         <Typography fontSize="header-24" fontWeight="bold" display="block">
-          판매차량 등록을 원하시나요?
+          {title}
         </Typography>
-        <Typography fontSize="body-14" color="greyScale-5" lineHeight="150%">
-          판매차량 등록 문의는 딜러 등록을 완료한 후에 가능합니다.
-        </Typography>
+        {subtitle && (
+          <Typography fontSize="body-14" color="greyScale-5" lineHeight="150%">
+            {subtitle}
+          </Typography>
+        )}
       </Wrapper.Top>
       <Wrapper.Bottom
         css={css`
@@ -50,18 +63,17 @@ const MarketBanner = () => {
           align-items: center;
         `}
       >
-        <Link href="/inquiry/market">
-          <Button
-            variant="Black"
-            border="rounded"
-            suffix={<ArrowRight fill="white" />}
-          >
-            등록 문의하기
-          </Button>
-        </Link>
+        <Button
+          variant="Black"
+          border="rounded"
+          suffix={<ArrowRight fill="white" />}
+          onClick={moveToUrl}
+        >
+          {btnTitle}
+        </Button>
       </Wrapper.Bottom>
     </Wrapper>
   );
 };
 
-export default React.memo(MarketBanner);
+export default React.memo(Announcement);
