+++ apps/web/src/components/common/announcement/index.ts
@@ -1,0 +1,1 @@
+export { default } from './announcement';
