--- apps/web/src/components/market/marketBanner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './marketBanner';
