--- apps/web/src/components/partnership/banner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './banner';
